+++ gazon-service/src/main/java/com/gitlab/repository/ReviewRepository.java
@@ -4,6 +4,7 @@ import com.gitlab.model.Review;
 import lombok.NonNull;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -20,5 +21,6 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
     @Override
     @NonNull
     @EntityGraph(value = "Review.reviewImages")
+    @Query("SELECT r FROM Review r WHERE r.entityStatus = 'ACTIVE'")
     List<Review> findAll();
 }